backend/requirements.txt
pydantic
typing
selenium 
webdriver-manager 
bs4
faiss-cpu
sentence_transformers
numpy
dotenv
logfire==3.1.0
openai==1.67.0;
python-dotenv==1.0.1
supabase==2.11.0
fastapi==0.110.0
scikit-learn
pydantic-ai==0.0.46
html2text
uvicorn==0.27.1
deepseek==1.0.0